// app/generated/movie.ts
export interface MovieItem {
  id: number;
  title: string;
  year: number;
  genre: string;
  rating: string;
  starRating: string;
  runtime: string;
  cast: string;
  image: string;
}

export interface MovieRequest {
}

export interface MovieResponse {
  movies: MovieItem[];
}

function createBaseMovieItem(): MovieItem {
  return { id: 0, title: "", year: 0, genre: "", rating: "", starRating: "", runtime: "", cast: "", image: "" };
}

export const MovieItem: MessageFns<MovieItem> = {
  encode(message: MovieItem, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (message.id !== 0) {
      writer.uint32(8).int32(message.id);
    }
    if (message.title !== "") {
      writer.uint32(18).string(message.title);
    }
    if (message.year !== 0) {
      writer.uint32(24).int32(message.year);
    }
    if (message.genre !== "") {
      writer.uint32(34).string(message.genre);
    }
    if (message.rating !== "") {
      writer.uint32(42).string(message.rating);
    }
    if (message.starRating !== "") {
      writer.uint32(50).string(message.starRating);
    }
    if (message.runtime !== "") {
      writer.uint32(58).string(message.runtime);
    }
    if (message.cast !== "") {
      writer.uint32(66).string(message.cast);
    }
    if (message.image !== "") {
      writer.uint32(74).string(message.image);
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): MovieItem {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    let end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseMovieItem();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 8) {
            break;
          }

          message.id = reader.int32();
          continue;
        }
        case 2: {
          if (tag !== 18) {
            break;
          }

          message.title = reader.string();
          continue;
        }
        case 3: {
          if (tag !== 24) {
            break;
          }

          message.year = reader.int32();
          continue;
        }
        case 4: {
          if (tag !== 34) {
            break;
          }

          message.genre = reader.string();
          continue;
        }
        case 5: {
          if (tag !== 42) {
            break;
          }

          message.rating = reader.string();
          continue;
        }
        case 6: {
          if (tag !== 50) {
            break;
          }

          message.starRating = reader.string();
          continue;
        }
        case 7: {
          if (tag !== 58) {
            break;
          }

          message.runtime = reader.string();
          continue;
        }
        case 8: {
          if (tag !== 66) {
            break;
          }

          message.cast = reader.string();
          continue;
        }
        case 9: {
          if (tag !== 74) {
            break;
          }

          message.image = reader.string();
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): MovieItem {
    return {
      id: isSet(object.id) ? globalThis.Number(object.id) : 0,
      title: isSet(object.title) ? globalThis.String(object.title) : "",
      year: isSet(object.year) ? globalThis.Number(object.year) : 0,
      genre: isSet(object.genre) ? globalThis.String(object.genre) : "",
      rating: isSet(object.rating) ? globalThis.String(object.rating) : "",
      starRating: isSet(object.starRating) ? globalThis.String(object.starRating) : "",
      runtime: isSet(object.runtime) ? globalThis.String(object.runtime) : "",
      cast: isSet(object.cast) ? globalThis.String(object.cast) : "",
      image: isSet(object.image) ? globalThis.String(object.image) : "",
    };
  },

  toJSON(message: MovieItem): unknown {
    const obj: any = {};
    if (message.id !== 0) {
      obj.id = Math.round(message.id);
    }
    if (message.title !== "") {
      obj.title = message.title;
    }
    if (message.year !== 0) {
      obj.year = Math.round(message.year);
    }
    if (message.genre !== "") {
      obj.genre = message.genre;
    }
    if (message.rating !== "") {
      obj.rating = message.rating;
    }
    if (message.starRating !== "") {
      obj.starRating = message.starRating;
    }
    if (message.runtime !== "") {
      obj.runtime = message.runtime;
    }
    if (message.cast !== "") {
      obj.cast = message.cast;
    }
    if (message.image !== "") {
      obj.image = message.image;
    }
    return obj;
  },

  create<I extends Exact<DeepPartial<MovieItem>, I>>(base?: I): MovieItem {
    return MovieItem.fromPartial(base ?? ({} as any));
  },
  fromPartial<I extends Exact<DeepPartial<MovieItem>, I>>(object: I): MovieItem {
    const message = createBaseMovieItem();
    message.id = object.id ?? 0;
    message.title = object.title ?? "";
    message.year = object.year ?? 0;
    message.genre = object.genre ?? "";
    message.rating = object.rating ?? "";
    message.starRating = object.starRating ?? "";
    message.runtime = object.runtime ?? "";
    message.cast = object.cast ?? "";
    message.image = object.image ?? "";
    return message;
  },
};
// ...
type Builtin = Date | Function | Uint8Array | string | number | boolean | undefined;

export type DeepPartial<T> = T extends Builtin ? T
  : T extends globalThis.Array<infer U> ? globalThis.Array<DeepPartial<U>>
  : T extends ReadonlyArray<infer U> ? ReadonlyArray<DeepPartial<U>>
  : T extends {} ? { [K in keyof T]?: DeepPartial<T[K]> }
  : Partial<T>;

type KeysOfUnion<T> = T extends T ? keyof T : never;
export type Exact<P, I extends P> = P extends Builtin ? P
  : P & { [K in keyof P]: Exact<P[K], I[K]> } & { [K in Exclude<keyof I, KeysOfUnion<P>>]: never };

function isSet(value: any): boolean {
  return value !== null && value !== undefined;
}

export class GrpcWebError extends globalThis.Error {
  constructor(message: string, public code: grpc.Code, public metadata: grpc.Metadata) {
    super(message);
  }
}

export interface MessageFns<T> {
  encode(message: T, writer?: BinaryWriter): BinaryWriter;
  decode(input: BinaryReader | Uint8Array, length?: number): T;
  fromJSON(object: any): T;
  toJSON(message: T): unknown;
  create<I extends Exact<DeepPartial<T>, I>>(base?: I): T;
  fromPartial<I extends Exact<DeepPartial<T>, I>>(object: I): T;
}